requirements.txt
numpy>=1.19.0
python-chess>=0.31.2
torch>=1.5.1
pyaml==20.4.0
PyYAML==5.3.1
